Rename EclairApi into Eclair and move it to root package
